qol fixes: Update Makefile and add systemd unit
This removes the makefile packaging target, since it wasn't really all that
useful for distribution of the program. Additionally, this sets up an install
target to install g13d to the users' local directory.

This is kinda a hack -- really we should be installing g13d to $PREFIX/bin,
dropping a g13d unit somewhere, and then setting up appropriate permissions,
etc. It works for me, though.