udev: Match attributes correctly and trigger systemd unit
This starts to make g13d more automatic. Instead of requiring manual
intervention or automatic starts on system start, this will allow udevd to
automatically trigger g13d when a g13 is plugged in.

There's still a bug in g13d where if the device falls off the USB chain g13d
won't exit, though, but this is a good start.