g13gui: First rough draft
This represents a full day of work just to get the binding behaviors correct and
the profile behaviors correct. At this point, this should be possible to turn
into something useful for controlling g13d with.

Things left to do:

  - Setup a thread for communicating with g13d with
  - Setup a worker queue to send profile changes or updates to the g13d daemon

Once those are in there, we can consider this to be feature complete for 1.0.
Additional functionality, such as LED colors, drawing to the LCD, and things
like supporting dbus IPC so we can handle applets on the LCD will come later.