g13gui: Got a bit carried away, totally reworked things
So at this point, g13gui has quite a lot of functionality built into it now. We
can bind keys, we can unbind keys, we can upload whole profiles to g13d, we can
create and edit profiles, and we can even set LCD colors per-profile.

We're not loading the configuration from disk, sadly, since I had to rework
quite a lot of the UI infrastructure to get observer notifications to work with
GTK. Lots of simplifications in here, though, which reduces the complexity of
the code considerably.

  - Migrated all model related classes into the model module.
  - Migrated UI classes into the ui module.
  - Migrated observer-related stuff into observer.
  - Created the GtkObserver adaptor for GTK UI threading. This makes heavy use
    of GObject signals and queuing to dispatch to the UI thread.
  - Migrated all of the G13 buttons into their own class.
  - Renamed ButtonMenu into G13ButtonPopover.
  - Setup the profile combo box as its own class.
  - Created the profile popover so we can add/remove/edit profiles.